+++ Day20.c
@@ -1,6 +1,6 @@
 #include "Helpers.c"
 
-static int parseAlgorithm(const char *input, bool algorithm[512]) {
+static uint32_t parseAlgorithm(const char *input, bool algorithm[512]) {
     char algoString[513] = {0};
     char firstInputRow[101] = {0};
     int charsRead = 0;
@@ -11,32 +11,32 @@ static int parseAlgorithm(const char *input, bool algorithm[512]) {
         algorithm[i] = algoString[i] == '#';
     }
 
-    return strlen(firstInputRow);
+    return (uint32_t)strlen(firstInputRow);
 }
 
-static void parseInputImage(const char *input, int n, bool image[n][n]) {
+static void parseInputImage(const char *input, uint32_t n, bool image[n][n]) {
     input += 512 + 2; // Assume two new lines.
 
     char row[101] = {0};
     int charsRead = 0;
 
-    for (int y = 0; y < n; ++y) {
+    for (uint32_t y = 0; y < n; ++y) {
         assert(sscanf(input, "%100s\n%n", row, &charsRead) == 1);
         input += charsRead;
 
-        for (int x = 0; x < n; ++x) {
+        for (uint32_t x = 0; x < n; ++x) {
             image[y][x] = row[x] == '#' ? 1 : 0;
         }
     }
 }
 
-static int indexFromInput(int sx, int sy, int inSize, int size, const bool inputImage[size][size], bool infinityBit) {
-    int index = 0;
+static uint32_t indexFromInput(int sx, int sy, uint32_t inSize, uint32_t size, const bool inputImage[size][size], bool infinityBit) {
+    uint32_t index = 0;
 
     for (int y = sy - 1; y <= sy + 1; ++y) {
         for (int x = sx - 1; x <= sx + 1; ++x) {
             index <<= 1;
-            index |= (x >= 0 && y >= 0 && x < inSize && y < inSize)
+            index |= (x >= 0 && y >= 0 && (uint32_t)x < inSize && (uint32_t)y < inSize)
                          ? inputImage[y][x]
                          : infinityBit;
         }
@@ -45,10 +45,10 @@ static int indexFromInput(int sx, int sy, int inSize, int size, const bool input
     return index;
 }
 
-static void enhance(const bool algorithm[512], int inSize, const bool inputImage[inSize][inSize], int outSize, bool outputImage[outSize][outSize]) {
+static void enhance(const bool algorithm[512], uint32_t inSize, const bool inputImage[inSize][inSize], uint32_t outSize, bool outputImage[outSize][outSize]) {
     assert(outSize > inSize);
 
-    int sizeBytes = sizeof(bool) * outSize * outSize;
+    size_t sizeBytes = sizeof(bool) * outSize * outSize;
 
     bool input[outSize][outSize];
     bool output[outSize][outSize];
@@ -56,9 +56,9 @@ static void enhance(const bool algorithm[512], int inSize, const bool inputImage
 
     int infinityBit = 0;
 
-    for (int y = 0; y < inSize + 2; ++y) {
-        for (int x = 0; x < inSize + 2; ++x) {
-            int index = indexFromInput(x - 1, y - 1, inSize, inSize, inputImage, infinityBit);
+    for (int y = 0; y < (int)inSize + 2; ++y) {
+        for (int x = 0; x < (int)inSize + 2; ++x) {
+            uint32_t index = indexFromInput(x - 1, y - 1, inSize, inSize, inputImage, infinityBit);
             output[y][x] = algorithm[index];
         }
     }
@@ -68,9 +68,9 @@ static void enhance(const bool algorithm[512], int inSize, const bool inputImage
 
         infinityBit = infinityBit ? algorithm[511] : algorithm[0]; // 0b111111111 = 511, 0b000000000 = 0.
 
-        for (int y = 0; y < inSize + 2; ++y) {
-            for (int x = 0; x < inSize + 2; ++x) {
-                int index = indexFromInput(x - 1, y - 1, inSize, outSize, input, infinityBit);
+        for (int y = 0; y < (int)inSize + 2; ++y) {
+            for (int x = 0; x < (int)inSize + 2; ++x) {
+                uint32_t index = indexFromInput(x - 1, y - 1, inSize, outSize, input, infinityBit);
                 output[y][x] = algorithm[index];
             }
         }
@@ -79,11 +79,11 @@ static void enhance(const bool algorithm[512], int inSize, const bool inputImage
     memcpy(outputImage, output, sizeBytes);
 }
 
-static int countLitPixels(int n, const bool image[n][n]) {
-    int count = 0;
+static uint32_t countLitPixels(uint32_t n, const bool image[n][n]) {
+    uint32_t count = 0;
 
-    for (int y = 0; y < n; ++y) {
-        for (int x = 0; x < n; ++x) {
+    for (uint32_t y = 0; y < n; ++y) {
+        for (uint32_t x = 0; x < n; ++x) {
             if (image[y][x]) {
                 ++count;
             }
@@ -93,12 +93,12 @@ static int countLitPixels(int n, const bool image[n][n]) {
     return count;
 }
 
-static int sizeAfterPasses(int n, int passes) {
+static uint32_t sizeAfterPasses(uint32_t n, uint32_t passes) {
     return n + 2 * passes;
 }
 
-static int partOne(const bool algorithm[512], int n, const bool inputImage[n][n]) {
-    int size = sizeAfterPasses(n, 2);
+static uint32_t partOne(const bool algorithm[512], uint32_t n, const bool inputImage[n][n]) {
+    uint32_t size = sizeAfterPasses(n, 2);
     bool outputImage[size][size];
 
     enhance(algorithm, n, inputImage, size, outputImage);
@@ -106,8 +106,8 @@ static int partOne(const bool algorithm[512], int n, const bool inputImage[n][n]
     return countLitPixels(size, outputImage);
 }
 
-static int partTwo(const bool algorithm[512], int n, const bool inputImage[n][n]) {
-    int size = sizeAfterPasses(n, 50);
+static uint32_t partTwo(const bool algorithm[512], uint32_t n, const bool inputImage[n][n]) {
+    uint32_t size = sizeAfterPasses(n, 50);
     bool outputImage[size][size];
     enhance(algorithm, n, inputImage, size, outputImage);
 
@@ -118,7 +118,7 @@ int main() {
     const char *input = Helpers_readInputFile(__FILE__);
 
     bool algorithm[512] = {0};
-    int n = parseAlgorithm(input, algorithm);
+    uint32_t n = parseAlgorithm(input, algorithm);
 
     bool inputImage[n][n];
     parseInputImage(input, n, inputImage);
